Renamed all services names
